feature: добавил базовую версию jpa репозиториев

+++ shared/src/main/java/backend/academy/shared/dto/TagLinkCount.java
@@ -1,3 +1,3 @@
 package backend.academy.shared.dto;
 
-public record TagLinkCount(String tag, int linksCount) {}
+public record TagLinkCount(String tag, long linksCount) {}
